render-deploy/requirements.txt
flask
flask-cors
google-generativeai
google-cloud-bigquery
pandas